创建店铺管理员

+++ src/main/java/com/github/star/a/admin/entity/AdminUser.java
@@ -15,7 +15,7 @@ public class AdminUser {
 
     String nickname;
 
-    Long isDeleted;
+    Integer isDeleted;
 
     Date createTime;
